Instalación admin template

+++ app/Http/Livewire/Admin/Category/CategoryList.php
@@ -33,9 +33,12 @@ class CategoryList extends Component
                 return $query->where('name', 'like', '%' . $this->search . '%')
                     ->orWhereHas('posts', function ($query) {
                         return $query->where('title', 'like', '%' . $this->search . '%');
+                    })
+                    ->orWhereHas('portfolios', function ($query) {
+                        return $query->where('title', 'like', '%' . $this->search . '%');
                     });
             })->OrderBy('id', 'desc')->paginate($this->perPage)->withQueryString(),
-        ])->extends('layouts.app')->section('content');
+        ])->extends('adminlte::page')->section('content');
     }
 
     // Confirmar eliminación
@@ -58,7 +61,7 @@ class CategoryList extends Component
         /* si esta relacionada con algun post mostrar un mensaje de error
         */
 
-        if ($category->posts->isEmpty()) {
+        if ($category->posts->isEmpty() && $category->portfolios->isEmpty()) {
             $category->delete();
             $this->dispatchBrowserEvent('toastr-delete', [
                 'title' => 'Categoría eliminada',
@@ -70,7 +73,7 @@ class CategoryList extends Component
             $this->dispatchBrowserEvent('toastr-no-delete', [
                 'title' => 'Categoría no eliminada',
                 'type' => 'error',
-                'message' => 'La categoría no puede ser eliminada porque esta relacionada con algun post',
+                'message' => 'La categoría no puede ser eliminada porque esta relacionada con algun post o portafolio',
                 'timer' => 3000,
             ]);
         }
